fix(fbvuln): Create fbvuln-manifest directories if needed
The release builds are built with INHERIT += "fbvuln-manifest" because we want to create the CVE manifest from the fbvuln-manifest.bbclass file. However at some point e2e-image stopped building this way and it was not discovered until release time.  This is because the build order of this particular target starts creating manifests before the deploy directory has been created.  Fix this by creating the directory if needed in fbvuln-manifest.bbclass.