Support Scarthgap CVE_STATUS
Scarthgap started a process of migrating from CVE_CHECK_IGNORE to
CVE_STATUS.  CVE_STATUS is a flagsdict rather than a simple variable,
which has keys of the CVE mapped to a status string.

Whinlatter started emitting a warning for deprecation of
CVE_CHECK_IGNORE.  Rework the use of CVE_CHECK_IGNORE and
CVE_CHECK_WHITELIST to both be backward compatible and without
the deprecation warnings.

```
WARNING: CVE_CHECK_IGNORE is deprecated in favor of CVE_STATUS
```